requirements.txt
torch
torchaudio
torchvision
librosa
matplotlib
numpy
pandas
scikit-learn
tqdm
opencv-python
PyYAML
seaborn
hydra-core
omegaconf